Refactor Home Carousel
